+++ docs/components/homePage/HomePageTitle.tsx
@@ -2,6 +2,7 @@ import { Title, createStyles } from '@mantine/core';
 
 const useStyles = createStyles((theme) => ({
   root: {
+    textIndent: '-0.1em',
     marginBottom: '.75em',
     color: theme.colorScheme === 'dark' ? theme.colors.dark[0] : theme.colors.gray[8],
     '@media (min-width: 320px)': {
@@ -37,11 +38,11 @@ export default function HomePageTitle() {
   const { classes } = useStyles();
   return (
     <Title className={classes.root} order={2}>
-      A table component
+      The <span className={classes.gradientText}>table component</span>
       <br />
       for your Mantine
       <br />
-      <span className={classes.gradientText}>data-rich applications.</span>
+      data-rich applications
     </Title>
   );
 }